examples/zero-code-examples/pydantic-ai/requirements.txt
pydantic-ai>=0.2.0
logfire-api>=2.0.0

opentelemetry-sdk>=1.36.0
opentelemetry-exporter-otlp-proto-http>=1.36.0
python-dotenv>=1.0.0
